pyproject.toml
[tool.poetry]
name = "recruitment"
version = "0.1.0"
description = ""
authors = ["willembspy <146359067+willembspy@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
faker = "^30.3.0"
pandas = "^2.2.3"
openai = "^1.51.2"
pinecone = "^5.3.1"
django = "^5.1.2"
asgiref = "^3.8.1"
django-cors-headers = "^4.5.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
pyjwt = "^2.9.0"
pytz = "^2024.2"
sqlparse = "^0.5.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
npm = "^0.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
